Remove nova.log usage from nova.rpc.
Part of blueprint common-rpc.

This patch removes the usage of nova.log from nova.rpc.  This is
necessary before nova.rpc can be moved to openstack-common.  The rpc
logging was not using any of the extra features provided in nova.log as
far as I can tell, so the changes are very simple.

This is the final dependency.  After this change, the rpc code
can be moved to openstack-common without any changes.

Change-Id: I17ca4b691bb8d4f18e618632a2381dd12d31852a