Add get_console_topic() to the compute rpcapi.
Part of bug 1006467.

This patch adds get_console_topic to the compute rpcapi.  This method is
used by the console API.  Honestly, the method seems a bit pointless,
but changing it further would be outside the scope of this patch.  This
is simply to get the code updated to handle rpc api versioning properly.

Change-Id: I738be88a52fede0739a73283b6e2640072f0f72b