Use explicit arguments in compute manager run_instance.
Use explicit arguments in functions in the run_instance code path of the
compute manager.  I think methods of the rpc API are better this way as
we have to be careful when changing these functions so we handle
versioning and backwards compatibility properly.  Spelling out all of
the arguments like this makes it easier to catch when things change and
to make sure they are changed properly.

This is in preparation of changing the instance_uuid to an instance
dict, instead.

Part of blueprint no-db-messaging.

Change-Id: I1e141b471040b0fd4833d4e08d44166a9d12bcad