Add V2 rpc api for cert
Add support for the 2.0 cert rpc API.  This commit retains compatibility
with the older 1.X API to allow continuous deployment without any downtime.

The v2 rev of this API is really simple.  It removes the
get_backdoor_port() method, which was moved to the base rpc API common
to all services in havana.

UpgradeImpact - Deployments doing continuous deployment should follow this
process to upgrade without any downtime with the consoleauth service:

1) Set [upgrade_levels] cert=havana in your config.

2) Upgrade to this commit.

3) Once everything has been upgraded, remove the entry in
[upgrade_levels] so that all clients start sending the new 2.0 messages.

Part of blueprint rpc-major-version-updates-icehouse

Change-Id: Icab0d5f380f43c454c166fa3479dacda74d89a02