Remove 2 uses of old _instance_update()
Remove 2 uses of the older _instance_update() method on the compute manager.
This method uses the db proxy method on the conductor which we are phasing out
in favor of using the Instance object.  These two cases were places where the
code already had an Instance object and just needed to do a new style update.

Part of blueprint compute-manager-objects

Change-Id: I01296602dcd108949ac73927b84416e307fb08ad